Essaie d'affichage

+++ noded.cpp
@@ -1,6 +1,6 @@
 #include "noded.h"
 
-NodeD::NodeD(unsigned int id, QPointF point, QString name): Node{id, point}, d_name{name}
+NodeD::NodeD(long long id, QPointF point, QString name): Node{id, point}, d_name{name}
 {}
 
 NodeD::NodeD(Node n, QString name): Node{n}, d_name{name}
